Refactor module internals
