+++ src/authorize/index.js
@@ -1,14 +1,9 @@
-const routes = require('../config/routes');
-const WTError = require('../error');
 const request = require('../request');
+const routes = require('../config/routes');
 
-module.exports = async function authorize() {
-  try {
-    return await request.send(routes.authorize);
-  } catch (error) {
-    throw new WTError(
-      'The authorization call failed and no usable JWT could be found there. Please check your API Key.',
-      error
-    );
-  }
-};
+const authorize = require('./authorize');
+
+module.exports = authorize({
+  request,
+  routes
+});
