+++ src/items/index.js
@@ -1,0 +1,7 @@
+const { addItems, completeFileUpload, uploadFile } = require('./actions');
+
+module.exports = {
+  addItems,
+  uploadFile,
+  completeFileUpload
+};
