+++ src/transfer/actions/create-and-upload.js
@@ -1,0 +1,5 @@
+module.exports = function() {
+  return async function createAndUpload() {
+    throw new Error('Not implemented yet');
+  };
+};
