+++ src/transfer/actions/index.js
@@ -1,0 +1,7 @@
+const request = require('../../request');
+const routes = require('../../config/routes');
+
+module.exports = {
+  create: require('./create')({ request, routes }),
+  createAndUpload: require('./create-and-upload')({ request, routes })
+};
